Add parentheses to demarcate optional portions
Checked with Oremus: they render fine.
A few entries in BCP lack commas separating off the portions in parentheses.
Amusingly, Christmas-I and Christmas-II have the same entry for all three years;
Years A & B with commas, Year C without!
Also add in a few letters for partial verses that didn't make it in the last
commit.
AND, switch which Psalm is the default for Friday in Easter week.